+++ robocorp-code/vscode-client/src/extension.ts
@@ -132,6 +132,7 @@ import {
     ROBOCORP_OPEN_LOCATORS_JSON,
     ROBOCORP_OPEN_ROBOT_CONDA_TREE_SELECTION,
     ROBOCORP_CONVERT_PROJECT,
+    ROBOCORP_NEW_ROBOCORP_INSPECTOR_WEB_RECORDER,
 } from "./robocorpCommands";
 import { installPythonInterpreterCheck } from "./pythonExtIntegration";
 import { refreshCloudTreeView } from "./viewsRobocorp";
@@ -483,9 +484,18 @@ function registerRobocorpCodeCommands(C: CommandRegistry) {
     C.register(ROBOCORP_EDIT_ROBOCORP_INSPECTOR_LOCATOR, (locator?: LocatorEntry) =>
         inspector.openRobocorpInspector(undefined, locator)
     );
-    C.register(ROBOCORP_NEW_ROBOCORP_INSPECTOR_BROWSER, () => inspector.openRobocorpInspector("browser"));
-    C.register(ROBOCORP_NEW_ROBOCORP_INSPECTOR_IMAGE, () => inspector.openRobocorpInspector("image"));
-    C.register(ROBOCORP_NEW_ROBOCORP_INSPECTOR_WINDOWS, () => inspector.openRobocorpInspector("windows"));
+    C.register(ROBOCORP_NEW_ROBOCORP_INSPECTOR_BROWSER, () =>
+        inspector.openRobocorpInspector(inspector.InspectorType.Browser)
+    );
+    C.register(ROBOCORP_NEW_ROBOCORP_INSPECTOR_IMAGE, () =>
+        inspector.openRobocorpInspector(inspector.InspectorType.Image)
+    );
+    C.register(ROBOCORP_NEW_ROBOCORP_INSPECTOR_WINDOWS, () =>
+        inspector.openRobocorpInspector(inspector.InspectorType.Windows)
+    );
+    C.register(ROBOCORP_NEW_ROBOCORP_INSPECTOR_WEB_RECORDER, () =>
+        inspector.openRobocorpInspector(inspector.InspectorType.WebRecorder)
+    );
     C.register(ROBOCORP_COPY_LOCATOR_TO_CLIPBOARD_INTERNAL, (locator?: LocatorEntry) =>
         copySelectedToClipboard(locator)
     );